+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## [Unreleased]
 
+- use shadowenv for configuring terraform backend secrets
+- use GitLab http backend for terraform state storage whenever configuring both terraform and GitLab CI
+
 ## [2.1.0] - 2025-04-29
 
 - Terraform generator updates to remove the old cloudfoundy-community provider and reduce the need for cloud.gov service accounts