+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## [Unreleased]
 
+- Prevent non-compliant hostnames by replacing underscores with dashes
 - use shadowenv for configuring terraform backend secrets
 - use GitLab http backend for terraform state storage whenever configuring both terraform and GitLab CI